rearange main.ts and refactoring, renaming of files and dir
