+++ src/app/app.module.ts
@@ -3,18 +3,17 @@ import { BrowserModule } from '@angular/platform-browser';
 import {HTTP_INTERCEPTORS, HttpClientModule} from '@angular/common/http';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { BannerComponent } from './banner/banner.component';
-import { MenuComponent } from './menu/menu.component';
-import { MainComponent } from './main/main.component';
-import { UserComponent } from './user/user.component';
+import { BannerComponent } from './customComponents/banner/banner.component';
+import { MenuComponent } from './customComponents/menu/menu.component';
+import { ChartComponent } from './chart/chart.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { AuthService } from './services/auth.service';
 import { AuthInterceptor } from './auth.interceptor';
-import { LegalMentionsComponent } from './legal-mentions/legal-mentions.component';
-import { HomeComponent } from './home/home.component';
-import { GraphLegendComponent } from './main/graph-legend/graph-legend.component';
+import { LegalMentionsComponent } from './pages/legal-mentions/legal-mentions.component';
+import { HomeComponent } from './pages/home/home.component';
 import { ToggleButtonComponent } from './customComponents/toggle-button/toggle-button.component';
-import { ResetPasswordComponent } from './reset-password/reset-password.component';
+import { ResetPasswordComponent } from './pages/reset-password/reset-password.component';
+import { LineChartComponent } from './chart/line-chart/line-chart.component';
 
 
 @NgModule({
@@ -22,13 +21,12 @@ import { ResetPasswordComponent } from './reset-password/reset-password.componen
     AppComponent,
     BannerComponent,
     MenuComponent,
-    MainComponent,
-    UserComponent,
+    ChartComponent,
     LegalMentionsComponent,
     HomeComponent,
-    GraphLegendComponent,
     ToggleButtonComponent,
-    ResetPasswordComponent
+    ResetPasswordComponent,
+    LineChartComponent
   ],
   imports: [
     BrowserModule,
