+++ src/app/chart/chart.component.scss
@@ -1,0 +1,43 @@
+
+#toast {
+    visibility: hidden;
+    position: absolute;
+    width: 100%;
+    left: 0;
+    top: 0;
+    text-align: center;
+    padding: 5px;
+    font-weight: 600;
+    &.show {
+        visibility: visible;
+        -webkit-animation: fadein 0.3s, fadeout 0.5s 3.7s;
+        animation: fadein 0.3s, fadeout 0.5s 3.7s;
+    }
+}
+
+//#toast.show {
+//    visibility: visible;
+//    -webkit-animation: fadein 0.3s, fadeout 0.5s 3.7s;
+//    animation: fadein 0.3s, fadeout 0.5s 3.7s;
+//}
+
+/* Animations to fade the toast in and out */
+@-webkit-keyframes fadein {
+    from {opacity: 0;}
+    to {opacity: 1;}
+}
+  
+@keyframes fadein {
+    from {opacity: 0;}
+    to {opacity: 1;}
+}
+  
+@-webkit-keyframes fadeout {
+    from {opacity: 1;}
+    to {opacity: 0;}
+}
+  
+@keyframes fadeout {
+    from {opacity: 1}
+    to {opacity: 0}
+}
