+++ src/app/customComponents/banner/banner.component.spec.ts
