+++ src/app/customComponents/banner/banner.component.ts
