--- src/app/main/graph-legend/graph-legend.component.html
@@ -1,20 +1,0 @@
-<div id="filters" class="absolute right-0 p-2 flex flex-col mt-80">
-    <div>
-        <button class="btn-graph" id="zoomButton" title="Zoomer">
-            <svg width="25px" height="25px" viewBox="-0.1 0 31.874 31.874" xmlns="http://www.w3.org/2000/svg">
-                <g id="Group_23" data-name="Group 23" transform="translate(-1199.025 -182)">
-                  <circle id="Ellipse_8" data-name="Ellipse 8" cx="11" cy="11" r="11" transform="translate(1200.025 183)" fill="#e8f7f9" stroke="#333" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"/>
-                  <path id="Path_25" data-name="Path 25" d="M1211,187a7,7,0,0,0-7,7" fill="none" stroke="#333" stroke-miterlimit="10" stroke-width="2"/>
-                  <line id="Line_24" data-name="Line 24" x1="11.568" y1="11.568" transform="translate(1218.432 201.599)" fill="none" stroke="#333" stroke-miterlimit="10" stroke-width="2"/>
-                </g>
-              </svg>
-        </button>
-        <button class="btn-graph" id="global_data" title="Données du dernier jour">
-            <img src="../../assets/earth.svg" width="25px" height="25px" alt="Bouton moyenne globale" title="Données moyennes de tous les utilisateurs">
-        </button>
-    </div>
-    <div>
-        <button class="btn-graph" id="day" title="Données du dernier jour">Aujourd'hui</button>
-        <button class="btn-graph" id="all" title="Données depuis le début">Début</button>
-    </div>
-</div>
