--- src/app/main/graph-legend/graph-legend.component.scss
@@ -1,7 +1,0 @@
-.btn-graph {
-    @apply rounded-lg border-2 border-slate-400 p-2 m-1 mb-2 bg-slate-100 hover:bg-lime-200 active:bg-lime-300
-}
-
-.activated  {
-    @apply rounded-lg border-2 border-lime-400 p-2 m-1 mb-2 bg-lime-300 hover:bg-slate-200 active:bg-slate-300
-}
