--- src/app/main/graph-legend/graph-legend.component.ts
@@ -1,14 +1,0 @@
-import { Component } from '@angular/core';
-import * as d3 from 'd3';
-
-@Component({
-  selector: 'app-graph-legend',
-  templateUrl: './graph-legend.component.html',
-  styleUrls: ['./graph-legend.component.scss']
-})
-export class GraphLegendComponent {
-
-  constructor() {}
-
-  ngOnInit() {}
-}
