+++ src/app/pages/legal-mentions/legal-mentions.component.scss
