+++ src/app/pages/reset-password/reset-password.component.ts
@@ -1,8 +1,8 @@
 import { Component } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { ToastService, toastType } from '../toast.service';
-import { UserService } from '../user/services/user.service';
-import { AuthService } from '../services/auth.service';
+import { ToastService, toastType } from '../../services/toast.service';
+import { UserService } from '../../services/user.service';
+import { AuthService } from '../../services/auth.service';
 import { ActivatedRoute, ParamMap } from '@angular/router';
 
 @Component({
