--- src/app/user/user.component.html
@@ -1,1 +1,0 @@
-<p>user works!</p>
